Adding file "source\ScriptModules\ScriptModulesMisc.h".
Reason, for easier overview.

Also adding a few comments on the #inlcudes in ScriptModules.cpp
